+++ src/Debug/RouterCollector.php
@@ -12,6 +12,9 @@ use Yiisoft\Router\RouteCollectionInterface;
 use Yiisoft\Yii\Debug\Collector\CollectorTrait;
 use Yiisoft\Yii\Debug\Collector\SummaryCollectorInterface;
 
+/**
+ * @codeCoverageIgnore
+ */
 final class RouterCollector implements SummaryCollectorInterface
 {
     use CollectorTrait;